Cleanup filter and allow for dropped channels
Cleanup: sorting the channels in the filter file was residual from when reconstruction could have been done in this same file; it's no longer necessary.

Drop channels: now, the filter file will do an "all channels" analysis, and a "dropped channels", and save both. This means the user has to decide which to use in the future, but that's okay.